test: harden reliable PR detection refresh
Issue: #4106
